fix(condo): DOMA-5403 access check to canManage... Move check fn in utils/accessSchema
