+++ apps/condo/domains/banking/access/BankCostItem.js
@@ -4,7 +4,7 @@
 
 const { throwAuthenticationError } = require('@open-condo/keystone/apolloErrorFormatter')
 
-const { checkBankIntegrationsAccessRights } = require('./BankIntegrationAccessRight')
+const { checkBankIntegrationsAccessRights } = require('@condo/domains/banking/utils/accessSchema')
 
 const { BANK_INTEGRATION_IDS } = require('../constants')
 
@@ -19,10 +19,12 @@ async function canReadBankCostItems ({ authentication: { item: user }, context }
     return {}
 }
 
-async function canManageBankCostItems ({ authentication: { item: user }, originalInput, operation, itemId }) {
+async function canManageBankCostItems ({ authentication: { item: user }, originalInput, operation, itemId, context }) {
     if (!user) return throwAuthenticationError()
     if (user.deletedAt) return false
     if (user.isAdmin) return true
+
+    if (await checkBankIntegrationsAccessRights(context, user.id, [BANK_INTEGRATION_IDS.SBBOL])) return true
 
     return false
 }
