+++ apps/condo/domains/banking/access/BankIntegrationAccessRight.js
@@ -2,15 +2,17 @@
  * Generated by `createschema banking.BankIntegrationAccessRight 'integration:Relationship:BankIntegration:PROTECT; user:Relationship:User:CASCADE;'`
  */
 
+const { userIsOwner } = require('@open-condo/keystone/access')
 const { throwAuthenticationError } = require('@open-condo/keystone/apolloErrorFormatter')
 
 const { BankIntegrationAccessRight } = require('@condo/domains/banking/utils/serverSchema')
 
-async function canReadBankIntegrationAccessRights ({ authentication: { item: user } }) {
+async function canReadBankIntegrationAccessRights (args) {
+    const { authentication: { item: user } } = args
     if (!user) return throwAuthenticationError()
     if (user.deletedAt) return false
-
     if (user.isAdmin || user.isSupport) return {}
+    if (userIsOwner(args)) return {}
 
     return false
 }
@@ -22,25 +24,7 @@ async function canManageBankIntegrationAccessRights ({ authentication: { item: u
     return !!(user.isSupport || user.isAdmin)
 }
 
-async function checkBankIntegrationsAccessRights (context, userId, integrationIds) {
-    if (!context) return false
-    if (!userId) return false
-    if (!Array.isArray(integrationIds) || !integrationIds.length || !integrationIds.every(Boolean)) return false
-
-    const rights = await BankIntegrationAccessRight.getAll(context, {
-        integration: { id_in: integrationIds },
-        user: { id: userId },
-        deletedAt: null,
-    }, { first: 100 })
-
-    const permittedIntegrations = new Set(rights.map(right => right.integration))
-    const nonPermittedIntegrations = integrationIds.filter(id => !permittedIntegrations.has(id))
-
-    return nonPermittedIntegrations.length === 0
-}
-
 module.exports = {
     canReadBankIntegrationAccessRights,
     canManageBankIntegrationAccessRights,
-    checkBankIntegrationsAccessRights,
 }
