+++ apps/condo/domains/banking/access/BankTransaction.js
@@ -1,15 +1,18 @@
 /**
  * Generated by `createschema banking.BankTransaction 'account:Relationship:BankAccount:CASCADE; contractorAccount:Relationship:BankContractorAccount:CASCADE; costItem?:Relationship:BankCostItem:SET_NULL; organization:Relationship:Organization:CASCADE; number:Text; date:CalendarDay; amount:Decimal; purpose:Text; dateWithdrawed:CalendarDay; dateReceived:CalendarDay; meta:Json; importId:Text; importRemoteSystem:Text;'`
  */
-const { throwAuthenticationError } = require('@open-condo/keystone/apolloErrorFormatter')
+const get = require('lodash/get')
 
+const { throwAuthenticationError } = require('@open-condo/keystone/apolloErrorFormatter')
+const { getById } = require('@open-condo/keystone/schema')
+
+const { checkBankIntegrationsAccessRights } = require('@condo/domains/banking/utils/accessSchema')
 const { canManageBankEntityWithOrganization } = require('@condo/domains/banking/utils/accessSchema')
 const {
     queryOrganizationEmployeeFor,
     queryOrganizationEmployeeFromRelatedOrganizationFor,
 } = require('@condo/domains/organization/utils/accessSchema')
 
-const { checkBankIntegrationsAccessRights } = require('./BankIntegrationAccessRight')
 
 const { BANK_INTEGRATION_IDS } = require('../constants')
 
@@ -30,12 +33,23 @@ async function canReadBankTransactions ({ authentication: { item: user }, contex
 }
 
 async function canManageBankTransactions (args) {
-    const { authentication: { item: user } } = args
+    const { authentication: { item: user }, originalInput, operation, itemId, context } = args
 
     if (!user) return throwAuthenticationError()
     if (user.deletedAt) return false
     if (user.isAdmin) return true
 
+    if (await checkBankIntegrationsAccessRights(context, user.id, [BANK_INTEGRATION_IDS.SBBOL])) return true
+
+    let organizationId
+
+    if (operation === 'create') {
+        organizationId = get(originalInput, ['organization', 'connect', 'id'])
+    }
+    if (operation === 'update') {
+        const bankTransaction = await getById('BankTransaction', itemId)
+        organizationId = get(bankTransaction, ['organization'])
+    }
     return canManageBankEntityWithOrganization(args, 'canManageBankTransactions')
 }
